refactor(subscription): move message related functionality to new class
