+++ test/subscription.js
@@ -18,19 +18,8 @@
 
 var assert = require('assert');
 var common = require('@google-cloud/common');
-var delay = require('delay');
-var events = require('events');
 var extend = require('extend');
-var is = require('is');
 var proxyquire = require('proxyquire');
-var util = require('util');
-
-var FAKE_FREE_MEM = 168222720;
-var fakeOs = {
-  freemem: function() {
-    return FAKE_FREE_MEM;
-  },
-};
 
 var promisified = false;
 var fakeUtil = extend({}, common.util, {
@@ -44,17 +33,6 @@ var fakeUtil = extend({}, common.util, {
   },
 });
 
-function FakeConnectionPool() {
-  this.calledWith_ = [].slice.call(arguments);
-  events.EventEmitter.call(this);
-}
-
-util.inherits(FakeConnectionPool, events.EventEmitter);
-
-function FakeHistogram() {
-  this.calledWith_ = [].slice.call(arguments);
-}
-
 function FakeIAM() {
   this.calledWith_ = [].slice.call(arguments);
 }
@@ -63,10 +41,8 @@ function FakeSnapshot() {
   this.calledWith_ = [].slice.call(arguments);
 }
 
-var delayOverride = null;
-
-function fakeDelay(timeout) {
-  return (delayOverride || delay)(timeout);
+function FakeSubscriber() {
+  this.calledWith_ = [].slice.call(arguments);
 }
 
 describe('Subscription', function() {
@@ -88,12 +64,9 @@ describe('Subscription', function() {
       '@google-cloud/common': {
         util: fakeUtil,
       },
-      delay: fakeDelay,
-      os: fakeOs,
-      './connection-pool.js': FakeConnectionPool,
-      './histogram.js': FakeHistogram,
       './iam.js': FakeIAM,
       './snapshot.js': FakeSnapshot,
+      './subscriber.js': FakeSubscriber,
     });
   });
 
@@ -128,14 +101,6 @@ describe('Subscription', function() {
       subscription.request(done);
     });
 
-    it('should create a histogram instance', function() {
-      assert(subscription.histogram instanceof FakeHistogram);
-    });
-
-    it('should create a latency histogram', function() {
-      assert(subscription.latency_ instanceof FakeHistogram);
-    });
-
     it('should format the sub name', function() {
       var formattedName = 'a/b/c/d';
       var formatName = Subscription.formatName_;
@@ -151,50 +116,6 @@ describe('Subscription', function() {
 
       var subscription = new Subscription(PUBSUB, SUB_NAME);
       assert.strictEqual(subscription.name, formattedName);
-    });
-
-    it('should honor configuration settings', function() {
-      var options = {
-        maxConnections: 2,
-        flowControl: {
-          maxBytes: 5,
-          maxMessages: 10,
-        },
-      };
-
-      var subscription = new Subscription(PUBSUB, SUB_NAME, options);
-
-      assert.strictEqual(subscription.maxConnections, options.maxConnections);
-
-      assert.deepEqual(subscription.flowControl, {
-        maxBytes: options.flowControl.maxBytes,
-        maxMessages: options.flowControl.maxMessages,
-      });
-    });
-
-    it('should set sensible defaults', function() {
-      assert.strictEqual(subscription.ackDeadline, 10000);
-      assert.strictEqual(subscription.maxConnections, 5);
-      assert.strictEqual(subscription.userClosed_, false);
-      assert.strictEqual(subscription.messageListeners, 0);
-      assert.strictEqual(subscription.isOpen, false);
-
-      assert.deepEqual(subscription.flowControl, {
-        maxBytes: FAKE_FREE_MEM * 0.2,
-        maxMessages: 100,
-      });
-    });
-
-    it('should create an inventory object', function() {
-      assert(is.object(subscription.inventory_));
-      assert(is.array(subscription.inventory_.lease));
-      assert(is.array(subscription.inventory_.ack));
-      assert(is.array(subscription.inventory_.nack));
-      assert.strictEqual(subscription.inventory_.bytes, 0);
-    });
-
-    it('should inherit from EventEmitter', function() {
-      assert(subscription instanceof events.EventEmitter);
     });
 
     it('should make a create method if a topic is found', function(done) {
@@ -222,17 +143,12 @@ describe('Subscription', function() {
       assert.strictEqual(args[1], subscription.name);
     });
 
-    it('should listen for events', function() {
-      var called = false;
-      var listenForEvents = Subscription.prototype.listenForEvents_;
+    it('should inherit from Subscriber', function() {
+      var options = {};
+      var subscription = new Subscription(PUBSUB, SUB_NAME, options);
 
-      Subscription.prototype.listenForEvents_ = function() {
-        Subscription.prototype.listenForEvents_ = listenForEvents;
-        called = true;
-      };
-
-      new Subscription(PUBSUB, SUB_NAME);
-      assert(called);
+      assert(subscription instanceof FakeSubscriber);
+      assert(subscription.calledWith_[0], options);
     });
   });
 
@@ -286,446 +202,6 @@ describe('Subscription', function() {
     it('should format name when given a complete name', function() {
       var formattedName = Subscription.formatName_(PROJECT_ID, SUB_FULL_NAME);
       assert.equal(formattedName, SUB_FULL_NAME);
-    });
-  });
-
-  describe('ack_', function() {
-    var MESSAGE = {
-      ackId: 'abc',
-      received: 12345,
-      connectionId: 'def',
-    };
-
-    beforeEach(function() {
-      subscription.breakLease_ = fakeUtil.noop;
-      subscription.histogram.add = fakeUtil.noop;
-      subscription.acknowledge_ = function() {
-        return Promise.resolve();
-      };
-      subscription.setFlushTimeout_ = function() {
-        return Promise.resolve();
-      };
-    });
-
-    it('should add the time it took to ack to the histogram', function(done) {
-      var fakeNow = 12381832;
-      var now = global.Date.now;
-
-      global.Date.now = function() {
-        global.Date.now = now;
-        return fakeNow;
-      };
-
-      subscription.histogram.add = function(time) {
-        assert.strictEqual(time, fakeNow - MESSAGE.received);
-        done();
-      };
-
-      subscription.ack_(MESSAGE);
-    });
-
-    describe('with connection', function() {
-      beforeEach(function() {
-        subscription.isConnected_ = function() {
-          return true;
-        };
-      });
-
-      it('should acknowledge if there is a connection', function(done) {
-        subscription.acknowledge_ = function(ackId, connectionId) {
-          assert.strictEqual(ackId, MESSAGE.ackId);
-          assert.strictEqual(connectionId, MESSAGE.connectionId);
-          setImmediate(done);
-          return Promise.resolve();
-        };
-
-        subscription.ack_(MESSAGE);
-      });
-
-      it('should break the lease on the message', function(done) {
-        subscription.breakLease_ = function(message) {
-          assert.strictEqual(message, MESSAGE);
-          done();
-        };
-
-        subscription.ack_(MESSAGE);
-      });
-    });
-
-    describe('without connection', function() {
-      beforeEach(function() {
-        subscription.isConnected_ = function() {
-          return false;
-        };
-      });
-
-      it('should queue the message to be acked if no connection', function(done) {
-        subscription.setFlushTimeout_ = function() {
-          assert(subscription.inventory_.ack.indexOf(MESSAGE.ackId) > -1);
-          done();
-        };
-
-        subscription.ack_(MESSAGE);
-      });
-
-      it('should break the lease on the message', function(done) {
-        subscription.breakLease_ = function(message) {
-          assert.strictEqual(message, MESSAGE);
-          done();
-        };
-
-        subscription.ack_(MESSAGE);
-      });
-    });
-  });
-
-  describe('acknowledge_', function() {
-    var fakeAckIds = ['a', 'b', 'c'];
-
-    var batchSize = 3000;
-    var tooManyFakeAckIds = Array(batchSize * 2.5)
-      .fill('a')
-      .map(function(x, i) {
-        return x + i;
-      });
-    var expectedCalls = Math.ceil(tooManyFakeAckIds.length / batchSize);
-
-    describe('without streaming connection', function() {
-      beforeEach(function() {
-        subscription.isConnected_ = function() {
-          return false;
-        };
-      });
-
-      it('should make the correct request', function(done) {
-        var fakePromisified = {
-          call: function(context, config) {
-            assert.strictEqual(context, subscription);
-            assert.strictEqual(config.client, 'SubscriberClient');
-            assert.strictEqual(config.method, 'acknowledge');
-            assert.strictEqual(config.reqOpts.subscription, subscription.name);
-            assert.deepEqual(config.reqOpts.ackIds, fakeAckIds);
-
-            setImmediate(done);
-
-            return Promise.resolve();
-          },
-        };
-
-        fakeUtil.promisify = function(fn) {
-          assert.strictEqual(fn, subscription.request);
-          return fakePromisified;
-        };
-
-        subscription.on('error', done);
-        subscription.acknowledge_(fakeAckIds);
-      });
-
-      it('should batch requests if there are too many ackIds', function(done) {
-        var receivedCalls = 0;
-
-        var fakePromisified = {
-          call: function(context, config) {
-            var offset = receivedCalls * batchSize;
-            var expectedAckIds = tooManyFakeAckIds.slice(
-              offset,
-              offset + batchSize
-            );
-
-            assert.deepEqual(config.reqOpts.ackIds, expectedAckIds);
-
-            receivedCalls += 1;
-            if (receivedCalls === expectedCalls) {
-              setImmediate(done);
-            }
-
-            return Promise.resolve();
-          },
-        };
-
-        fakeUtil.promisify = function() {
-          return fakePromisified;
-        };
-
-        subscription.on('error', done);
-        subscription.acknowledge_(tooManyFakeAckIds);
-      });
-
-      it('should emit any request errors', function(done) {
-        var fakeError = new Error('err');
-        var fakePromisified = {
-          call: function() {
-            return Promise.reject(fakeError);
-          },
-        };
-
-        fakeUtil.promisify = function() {
-          return fakePromisified;
-        };
-
-        subscription.on('error', function(err) {
-          assert.strictEqual(err, fakeError);
-          done();
-        });
-
-        subscription.acknowledge_(fakeAckIds);
-      });
-    });
-
-    describe('with streaming connection', function() {
-      beforeEach(function() {
-        subscription.isConnected_ = function() {
-          return true;
-        };
-      });
-
-      it('should send the correct request', function(done) {
-        var fakeConnectionId = 'abc';
-
-        subscription.writeTo_ = function(connectionId, data) {
-          assert.strictEqual(connectionId, fakeConnectionId);
-          assert.deepEqual(data, {ackIds: fakeAckIds});
-          done();
-        };
-
-        subscription.acknowledge_(fakeAckIds, fakeConnectionId);
-      });
-
-      it('should batch requests if there are too many ackIds', function(done) {
-        var receivedCalls = 0;
-        var fakeConnectionId = 'abc';
-
-        subscription.writeTo_ = function(connectionId, data) {
-          assert.strictEqual(connectionId, fakeConnectionId);
-
-          var offset = receivedCalls * batchSize;
-          var expectedAckIds = tooManyFakeAckIds.slice(
-            offset,
-            offset + batchSize
-          );
-
-          assert.deepEqual(data, {ackIds: expectedAckIds});
-
-          if (++receivedCalls === expectedCalls) {
-            done();
-          }
-        };
-
-        subscription.acknowledge_(tooManyFakeAckIds, fakeConnectionId);
-      });
-
-      it('should emit an error when unable to get a conn', function(done) {
-        var error = new Error('err');
-
-        subscription.writeTo_ = function() {
-          return Promise.reject(error);
-        };
-
-        subscription.on('error', function(err) {
-          assert.strictEqual(err, error);
-          done();
-        });
-
-        subscription.acknowledge_(fakeAckIds);
-      });
-    });
-  });
-
-  describe('breakLease_', function() {
-    var MESSAGE = {
-      ackId: 'abc',
-      data: Buffer.from('hello'),
-      length: 5,
-    };
-
-    beforeEach(function() {
-      subscription.inventory_.lease.push(MESSAGE.ackId);
-      subscription.inventory_.bytes += MESSAGE.length;
-    });
-
-    it('should remove the message from the lease array', function() {
-      assert.strictEqual(subscription.inventory_.lease.length, 1);
-      assert.strictEqual(subscription.inventory_.bytes, MESSAGE.length);
-
-      subscription.breakLease_(MESSAGE);
-
-      assert.strictEqual(subscription.inventory_.lease.length, 0);
-      assert.strictEqual(subscription.inventory_.bytes, 0);
-    });
-
-    it('should noop for unknown messages', function() {
-      var message = {
-        ackId: 'def',
-        data: Buffer.from('world'),
-        length: 5,
-      };
-
-      subscription.breakLease_(message);
-
-      assert.strictEqual(subscription.inventory_.lease.length, 1);
-      assert.strictEqual(subscription.inventory_.bytes, 5);
-    });
-
-    describe('with connection pool', function() {
-      it('should resume receiving messages if paused', function(done) {
-        subscription.connectionPool = {
-          isPaused: true,
-          resume: done,
-        };
-
-        subscription.hasMaxMessages_ = function() {
-          return false;
-        };
-
-        subscription.breakLease_(MESSAGE);
-      });
-
-      it('should not resume if it is not paused', function() {
-        subscription.connectionPool = {
-          isPaused: false,
-          resume: function() {
-            throw new Error('Should not be called.');
-          },
-        };
-
-        subscription.hasMaxMessages_ = function() {
-          return false;
-        };
-
-        subscription.breakLease_(MESSAGE);
-      });
-
-      it('should not resume if the max message limit is hit', function() {
-        subscription.connectionPool = {
-          isPaused: true,
-          resume: function() {
-            throw new Error('Should not be called.');
-          },
-        };
-
-        subscription.hasMaxMessages_ = function() {
-          return true;
-        };
-
-        subscription.breakLease_(MESSAGE);
-      });
-    });
-
-    it('should quit auto-leasing if all leases are gone', function(done) {
-      subscription.leaseTimeoutHandle_ = setTimeout(done, 1);
-      subscription.breakLease_(MESSAGE);
-
-      assert.strictEqual(subscription.leaseTimeoutHandle_, null);
-      setImmediate(done);
-    });
-
-    it('should continue to auto-lease if leases exist', function(done) {
-      subscription.inventory_.lease.push(MESSAGE.ackId);
-      subscription.inventory_.lease.push('abcd');
-
-      subscription.leaseTimeoutHandle_ = setTimeout(done, 1);
-      subscription.breakLease_(MESSAGE);
-    });
-  });
-
-  describe('close', function() {
-    beforeEach(function() {
-      subscription.flushQueues_ = function() {
-        return Promise.resolve();
-      };
-
-      subscription.closeConnection_ = fakeUtil.noop;
-    });
-
-    it('should set the userClosed_ flag', function() {
-      subscription.close();
-
-      assert.strictEqual(subscription.userClosed_, true);
-    });
-
-    it('should dump the inventory', function() {
-      subscription.inventory_ = {
-        lease: [0, 1, 2],
-        bytes: 123,
-      };
-
-      subscription.close();
-
-      assert.deepEqual(subscription.inventory_, {
-        lease: [],
-        bytes: 0,
-      });
-    });
-
-    it('should stop auto-leasing', function(done) {
-      subscription.leaseTimeoutHandle_ = setTimeout(done, 1);
-      subscription.close();
-
-      assert.strictEqual(subscription.leaseTimeoutHandle_, null);
-      setImmediate(done);
-    });
-
-    it('should flush immediately', function(done) {
-      subscription.flushQueues_ = function() {
-        setImmediate(done);
-        return Promise.resolve();
-      };
-
-      subscription.close();
-    });
-
-    it('should call closeConnection_', function(done) {
-      subscription.closeConnection_ = function(callback) {
-        callback(); // the done fn
-      };
-
-      subscription.close(done);
-    });
-  });
-
-  describe('closeConnection_', function() {
-    afterEach(function() {
-      fakeUtil.noop = function() {};
-    });
-
-    it('should set isOpen to false', function() {
-      subscription.closeConnection_();
-      assert.strictEqual(subscription.isOpen, false);
-    });
-
-    describe('with connection pool', function() {
-      beforeEach(function() {
-        subscription.connectionPool = {
-          close: function(callback) {
-            setImmediate(callback); // the done fn
-          },
-        };
-      });
-
-      it('should call close on the connection pool', function(done) {
-        subscription.closeConnection_(done);
-        assert.strictEqual(subscription.connectionPool, null);
-      });
-
-      it('should use a noop when callback is absent', function(done) {
-        fakeUtil.noop = done;
-        subscription.closeConnection_();
-        assert.strictEqual(subscription.connectionPool, null);
-      });
-    });
-
-    describe('without connection pool', function() {
-      beforeEach(function() {
-        subscription.connectionPool = null;
-      });
-
-      it('should exec the callback if one is passed in', function(done) {
-        subscription.closeConnection_(done);
-      });
-
-      it('should optionally accept a callback', function() {
-        subscription.closeConnection_();
-      });
     });
   });
 
@@ -810,6 +286,11 @@ describe('Subscription', function() {
   });
 
   describe('delete', function() {
+    beforeEach(function() {
+      subscription.removeAllListeners = fakeUtil.noop;
+      subscription.close = fakeUtil.noop;
+    });
+
     it('should make the correct request', function(done) {
       subscription.request = function(config) {
         assert.strictEqual(config.client, 'SubscriberClient');
@@ -962,58 +443,6 @@ describe('Subscription', function() {
         assert.strictEqual(err, error);
         assert.strictEqual(exists, undefined);
         done();
-      });
-    });
-  });
-
-  describe('flushQueues_', function() {
-    it('should cancel any pending flushes', function() {
-      var canceled = false;
-      var fakeHandle = {
-        cancel: function() {
-          canceled = true;
-        },
-      };
-
-      subscription.flushTimeoutHandle_ = fakeHandle;
-      subscription.flushQueues_();
-
-      assert.strictEqual(subscription.flushTimeoutHandle_, null);
-      assert.strictEqual(canceled, true);
-    });
-
-    it('should do nothing if theres nothing to ack/nack', function() {
-      subscription.acknowledge_ = subscription.modifyAckDeadline_ = function() {
-        throw new Error('Should not be called.');
-      };
-
-      return subscription.flushQueues_();
-    });
-
-    it('should send any pending acks', function() {
-      var fakeAckIds = (subscription.inventory_.ack = ['abc', 'def']);
-
-      subscription.acknowledge_ = function(ackIds) {
-        assert.strictEqual(ackIds, fakeAckIds);
-        return Promise.resolve();
-      };
-
-      return subscription.flushQueues_().then(function() {
-        assert.strictEqual(subscription.inventory_.ack.length, 0);
-      });
-    });
-
-    it('should send any pending nacks', function() {
-      var fakeAckIds = (subscription.inventory_.nack = ['ghi', 'jkl']);
-
-      subscription.modifyAckDeadline_ = function(ackIds, deadline) {
-        assert.strictEqual(ackIds, fakeAckIds);
-        assert.strictEqual(deadline, 0);
-        return Promise.resolve();
-      };
-
-      return subscription.flushQueues_().then(function() {
-        assert.strictEqual(subscription.inventory_.nack.length, 0);
       });
     });
   });
@@ -1195,340 +624,6 @@ describe('Subscription', function() {
     });
   });
 
-  describe('isConnected_', function() {
-    it('should return false if there is no pool', function() {
-      subscription.connectionPool = null;
-      assert.strictEqual(subscription.isConnected_(), false);
-    });
-
-    it('should return false if the pool says its connected', function() {
-      subscription.connectionPool = {
-        isConnected: function() {
-          return false;
-        },
-      };
-
-      assert.strictEqual(subscription.isConnected_(), false);
-    });
-
-    it('should return true if the pool says its connected', function() {
-      subscription.connectionPool = {
-        isConnected: function() {
-          return true;
-        },
-      };
-
-      assert.strictEqual(subscription.isConnected_(), true);
-    });
-  });
-
-  describe('hasMaxMessages_', function() {
-    it('should return true if the number of leases >= maxMessages', function() {
-      subscription.inventory_.lease = ['a', 'b', 'c'];
-      subscription.flowControl.maxMessages = 3;
-
-      assert(subscription.hasMaxMessages_());
-    });
-
-    it('should return true if bytes == maxBytes', function() {
-      subscription.inventory_.bytes = 1000;
-      subscription.flowControl.maxBytes = 1000;
-
-      assert(subscription.hasMaxMessages_());
-    });
-
-    it('should return false if neither condition is met', function() {
-      subscription.inventory_.lease = ['a', 'b'];
-      subscription.flowControl.maxMessages = 3;
-
-      subscription.inventory_.bytes = 900;
-      subscription.flowControl.maxBytes = 1000;
-
-      assert.strictEqual(subscription.hasMaxMessages_(), false);
-    });
-  });
-
-  describe('leaseMessage_', function() {
-    var MESSAGE = {
-      ackId: 'abc',
-      connectionId: 'def',
-      data: Buffer.from('hello'),
-      length: 5,
-    };
-
-    beforeEach(function() {
-      subscription.setLeaseTimeout_ = fakeUtil.noop;
-      subscription.modifyAckDeadline_ = fakeUtil.noop;
-    });
-
-    it('should immediately modAck the message', function(done) {
-      subscription.modifyAckDeadline_ = function(ackId, deadline, connId) {
-        assert.strictEqual(ackId, MESSAGE.ackId);
-        assert.strictEqual(deadline, subscription.ackDeadline / 1000);
-        assert.strictEqual(connId, MESSAGE.connectionId);
-        done();
-      };
-
-      subscription.leaseMessage_(MESSAGE);
-    });
-
-    it('should add the ackId to the inventory', function() {
-      subscription.leaseMessage_(MESSAGE);
-      assert.deepEqual(subscription.inventory_.lease, [MESSAGE.ackId]);
-    });
-
-    it('should update the byte count', function() {
-      assert.strictEqual(subscription.inventory_.bytes, 0);
-      subscription.leaseMessage_(MESSAGE);
-      assert.strictEqual(subscription.inventory_.bytes, MESSAGE.length);
-    });
-
-    it('should begin auto-leasing', function(done) {
-      subscription.setLeaseTimeout_ = done;
-      subscription.leaseMessage_(MESSAGE);
-    });
-
-    it('should return the message', function() {
-      var message = subscription.leaseMessage_(MESSAGE);
-      assert.strictEqual(message, MESSAGE);
-    });
-  });
-
-  describe('listenForEvents_', function() {
-    beforeEach(function() {
-      subscription.openConnection_ = fakeUtil.noop;
-      subscription.closeConnection_ = fakeUtil.noop;
-    });
-
-    describe('on new listener', function() {
-      it('should increment messageListeners', function() {
-        assert.strictEqual(subscription.messageListeners, 0);
-        subscription.on('message', fakeUtil.noop);
-        assert.strictEqual(subscription.messageListeners, 1);
-      });
-
-      it('should ignore non-message events', function() {
-        subscription.on('data', fakeUtil.noop);
-        assert.strictEqual(subscription.messageListeners, 0);
-      });
-
-      it('should open a connection', function(done) {
-        subscription.openConnection_ = done;
-        subscription.on('message', fakeUtil.noop);
-      });
-
-      it('should set the userClosed_ flag to false', function() {
-        subscription.userClosed_ = true;
-        subscription.on('message', fakeUtil.noop);
-        assert.strictEqual(subscription.userClosed_, false);
-      });
-
-      it('should not open a connection when one exists', function() {
-        subscription.connectionPool = {};
-
-        subscription.openConnection_ = function() {
-          throw new Error('Should not be called.');
-        };
-
-        subscription.on('message', fakeUtil.noop);
-      });
-    });
-
-    describe('on remove listener', function() {
-      var noop = function() {};
-
-      it('should decrement messageListeners', function() {
-        subscription.on('message', fakeUtil.noop);
-        subscription.on('message', noop);
-        assert.strictEqual(subscription.messageListeners, 2);
-
-        subscription.removeListener('message', noop);
-        assert.strictEqual(subscription.messageListeners, 1);
-      });
-
-      it('should ignore non-message events', function() {
-        subscription.on('message', fakeUtil.noop);
-        subscription.on('message', noop);
-        assert.strictEqual(subscription.messageListeners, 2);
-
-        subscription.removeListener('data', noop);
-        assert.strictEqual(subscription.messageListeners, 2);
-      });
-
-      it('should close the connection when no listeners', function(done) {
-        subscription.closeConnection_ = done;
-
-        subscription.on('message', noop);
-        subscription.removeListener('message', noop);
-      });
-    });
-  });
-
-  describe('modifyAckDeadline_', function() {
-    var fakeAckIds = ['a', 'b', 'c'];
-    var fakeDeadline = 123;
-
-    var batchSize = 3000;
-    var tooManyFakeAckIds = Array(batchSize * 2.5)
-      .fill('a')
-      .map(function(x, i) {
-        return x + i;
-      });
-    var expectedCalls = Math.ceil(tooManyFakeAckIds.length / batchSize);
-
-    describe('without streaming connection', function() {
-      beforeEach(function() {
-        subscription.isConnected_ = function() {
-          return false;
-        };
-      });
-
-      it('should make the correct request', function(done) {
-        var fakePromisified = {
-          call: function(context, config) {
-            assert.strictEqual(context, subscription);
-            assert.strictEqual(config.client, 'SubscriberClient');
-            assert.strictEqual(config.method, 'modifyAckDeadline');
-            assert.strictEqual(config.reqOpts.subscription, subscription.name);
-            assert.strictEqual(config.reqOpts.ackDeadlineSeconds, fakeDeadline);
-            assert.deepEqual(config.reqOpts.ackIds, fakeAckIds);
-
-            setImmediate(done);
-
-            return Promise.resolve();
-          },
-        };
-
-        fakeUtil.promisify = function(fn) {
-          assert.strictEqual(fn, subscription.request);
-          return fakePromisified;
-        };
-
-        subscription.on('error', done);
-        subscription.modifyAckDeadline_(fakeAckIds, fakeDeadline);
-      });
-
-      it('should batch requests if there are too many ackIds', function(done) {
-        var receivedCalls = 0;
-
-        var fakePromisified = {
-          call: function(context, config) {
-            var offset = receivedCalls * batchSize;
-            var expectedAckIds = tooManyFakeAckIds.slice(
-              offset,
-              offset + batchSize
-            );
-
-            assert.strictEqual(config.reqOpts.ackDeadlineSeconds, fakeDeadline);
-            assert.deepEqual(config.reqOpts.ackIds, expectedAckIds);
-
-            receivedCalls += 1;
-            if (receivedCalls === expectedCalls) {
-              setImmediate(done);
-            }
-
-            return Promise.resolve();
-          },
-        };
-
-        fakeUtil.promisify = function() {
-          return fakePromisified;
-        };
-
-        subscription.on('error', done);
-        subscription.modifyAckDeadline_(tooManyFakeAckIds, fakeDeadline);
-      });
-
-      it('should emit any request errors', function(done) {
-        var fakeError = new Error('err');
-        var fakePromisified = {
-          call: function() {
-            return Promise.reject(fakeError);
-          },
-        };
-
-        fakeUtil.promisify = function() {
-          return fakePromisified;
-        };
-
-        subscription.on('error', function(err) {
-          assert.strictEqual(err, fakeError);
-          done();
-        });
-
-        subscription.modifyAckDeadline_(fakeAckIds, fakeDeadline);
-      });
-    });
-
-    describe('with streaming connection', function() {
-      beforeEach(function() {
-        subscription.isConnected_ = function() {
-          return true;
-        };
-      });
-
-      it('should send the correct request', function(done) {
-        var expectedDeadlines = Array(fakeAckIds.length).fill(fakeDeadline);
-        var fakeConnId = 'abc';
-
-        subscription.writeTo_ = function(connectionId, data) {
-          assert.strictEqual(connectionId, fakeConnId);
-          assert.deepEqual(data.modifyDeadlineAckIds, fakeAckIds);
-          assert.deepEqual(data.modifyDeadlineSeconds, expectedDeadlines);
-          done();
-        };
-
-        subscription.modifyAckDeadline_(fakeAckIds, fakeDeadline, fakeConnId);
-      });
-
-      it('should batch requests if there are too many ackIds', function(done) {
-        var receivedCalls = 0;
-        var fakeConnId = 'abc';
-
-        subscription.writeTo_ = function(connectionId, data) {
-          assert.strictEqual(connectionId, fakeConnId);
-
-          var offset = receivedCalls * batchSize;
-          var expectedAckIds = tooManyFakeAckIds.slice(
-            offset,
-            offset + batchSize
-          );
-          var expectedDeadlines = Array(expectedAckIds.length).fill(
-            fakeDeadline
-          );
-
-          assert.deepEqual(data.modifyDeadlineAckIds, expectedAckIds);
-          assert.deepEqual(data.modifyDeadlineSeconds, expectedDeadlines);
-
-          if (++receivedCalls === expectedCalls) {
-            done();
-          }
-        };
-
-        subscription.modifyAckDeadline_(
-          tooManyFakeAckIds,
-          fakeDeadline,
-          fakeConnId
-        );
-      });
-
-      it('should emit an error when unable to get a conn', function(done) {
-        var error = new Error('err');
-
-        subscription.writeTo_ = function() {
-          return Promise.reject(error);
-        };
-
-        subscription.on('error', function(err) {
-          assert.strictEqual(err, error);
-          done();
-        });
-
-        subscription.modifyAckDeadline_(fakeAckIds, fakeDeadline);
-      });
-    });
-  });
-
   describe('modifyPushConfig', function() {
     var fakeConfig = {};
 
@@ -1555,256 +650,6 @@ describe('Subscription', function() {
       };
 
       subscription.modifyPushConfig(fakeConfig, gaxOpts, done);
-    });
-  });
-
-  describe('nack_', function() {
-    var MESSAGE = {
-      ackId: 'abc',
-      connectionId: 'def',
-    };
-
-    beforeEach(function() {
-      subscription.breakLease_ = fakeUtil.noop;
-      subscription.modifyAckDeadline_ = function() {
-        return Promise.resolve();
-      };
-      subscription.setFlushTimeout_ = function() {
-        return Promise.resolve();
-      };
-    });
-
-    describe('with connection', function() {
-      beforeEach(function() {
-        subscription.isConnected_ = function() {
-          return true;
-        };
-      });
-
-      it('should nack if there is a connection', function(done) {
-        subscription.modifyAckDeadline_ = function(ackId, deadline, connId) {
-          assert.strictEqual(ackId, MESSAGE.ackId);
-          assert.strictEqual(deadline, 0);
-          assert.strictEqual(connId, MESSAGE.connectionId);
-          setImmediate(done);
-          return Promise.resolve();
-        };
-
-        subscription.nack_(MESSAGE);
-      });
-
-      it('should break the lease on the message', function(done) {
-        subscription.breakLease_ = function(message) {
-          assert.strictEqual(message, MESSAGE);
-          done();
-        };
-
-        subscription.nack_(MESSAGE);
-      });
-    });
-
-    describe('without connection', function() {
-      beforeEach(function() {
-        subscription.isConnected_ = function() {
-          return false;
-        };
-      });
-
-      it('should queue the message to be nacked if no conn', function(done) {
-        subscription.setFlushTimeout_ = function() {
-          assert(subscription.inventory_.nack.indexOf(MESSAGE.ackId) > -1);
-          setImmediate(done);
-          return Promise.resolve();
-        };
-
-        subscription.nack_(MESSAGE);
-      });
-
-      it('should break the lease on the message', function(done) {
-        subscription.breakLease_ = function(message) {
-          assert.strictEqual(message, MESSAGE);
-          done();
-        };
-
-        subscription.nack_(MESSAGE);
-      });
-    });
-  });
-
-  describe('openConnection_', function() {
-    it('should create a ConnectionPool instance', function() {
-      subscription.openConnection_();
-      assert(subscription.connectionPool instanceof FakeConnectionPool);
-
-      var args = subscription.connectionPool.calledWith_;
-      assert.strictEqual(args[0], subscription);
-    });
-
-    it('should emit pool errors', function(done) {
-      var error = new Error('err');
-
-      subscription.on('error', function(err) {
-        assert.strictEqual(err, error);
-        done();
-      });
-
-      subscription.openConnection_();
-      subscription.connectionPool.emit('error', error);
-    });
-
-    it('should set isOpen to true', function() {
-      subscription.openConnection_();
-      assert.strictEqual(subscription.isOpen, true);
-    });
-
-    it('should lease & emit messages from pool', function(done) {
-      var message = {};
-      var leasedMessage = {};
-
-      subscription.leaseMessage_ = function(message_) {
-        assert.strictEqual(message_, message);
-        return leasedMessage;
-      };
-
-      subscription.on('message', function(message) {
-        assert.strictEqual(message, leasedMessage);
-        done();
-      });
-
-      subscription.openConnection_();
-      subscription.connectionPool.emit('message', message);
-    });
-
-    it('should pause the pool if sub is at max messages', function(done) {
-      var message = {nack: fakeUtil.noop};
-      var leasedMessage = {};
-
-      subscription.leaseMessage_ = function() {
-        return leasedMessage;
-      };
-
-      subscription.hasMaxMessages_ = function() {
-        return true;
-      };
-
-      subscription.openConnection_();
-      subscription.connectionPool.isPaused = false;
-      subscription.connectionPool.pause = done;
-      subscription.connectionPool.emit('message', message);
-    });
-
-    it('should not re-pause the pool', function(done) {
-      var message = {nack: fakeUtil.noop};
-      var leasedMessage = {};
-
-      subscription.leaseMessage_ = function() {
-        return leasedMessage;
-      };
-
-      subscription.hasMaxMessages_ = function() {
-        return true;
-      };
-
-      subscription.openConnection_();
-      subscription.connectionPool.isPaused = true;
-
-      subscription.connectionPool.pause = function() {
-        done(new Error('Should not have been called.'));
-      };
-
-      subscription.connectionPool.emit('message', message);
-      done();
-    });
-
-    it('should flush the queue when connected', function(done) {
-      subscription.flushQueues_ = done;
-
-      subscription.openConnection_();
-      subscription.connectionPool.emit('connected');
-    });
-  });
-
-  describe('renewLeases_', function() {
-    beforeEach(function() {
-      subscription.modifyAckDeadline_ = function() {
-        return Promise.resolve();
-      };
-    });
-
-    var fakeDeadline = 9999;
-    var fakeAckIds = ['abc', 'def'];
-
-    beforeEach(function() {
-      subscription.inventory_.lease = fakeAckIds;
-      subscription.setLeaseTimeout_ = fakeUtil.noop;
-
-      subscription.histogram.percentile = function() {
-        return fakeDeadline;
-      };
-    });
-
-    it('should clean up the old timeout handle', function() {
-      var fakeHandle = 123;
-      var clearTimeoutCalled = false;
-      var _clearTimeout = global.clearTimeout;
-
-      global.clearTimeout = function(handle) {
-        assert.strictEqual(handle, fakeHandle);
-        clearTimeoutCalled = true;
-      };
-
-      subscription.leaseTimeoutHandle_ = fakeHandle;
-      subscription.renewLeases_();
-
-      assert.strictEqual(subscription.leaseTimeoutHandle_, null);
-      assert.strictEqual(clearTimeoutCalled, true);
-
-      global.clearTimeout = _clearTimeout;
-    });
-
-    it('should update the ackDeadline', function() {
-      subscription.request = subscription.setLeaseTimeout_ = fakeUtil.noop;
-
-      subscription.histogram.percentile = function(percent) {
-        assert.strictEqual(percent, 99);
-        return fakeDeadline;
-      };
-
-      subscription.renewLeases_();
-      assert.strictEqual(subscription.ackDeadline, fakeDeadline);
-    });
-
-    it('should set the auto-lease timeout', function(done) {
-      subscription.request = fakeUtil.noop;
-      subscription.setLeaseTimeout_ = done;
-      subscription.renewLeases_();
-    });
-
-    it('should not renew leases if inventory is empty', function() {
-      subscription.modifyAckDeadline_ = function() {
-        throw new Error('Should not have been called.');
-      };
-
-      subscription.inventory_.lease = [];
-      subscription.renewLeases_();
-    });
-
-    it('should modAck the leased messages', function(done) {
-      subscription.modifyAckDeadline_ = function(ackIds, deadline) {
-        assert.deepEqual(ackIds, fakeAckIds);
-        assert.strictEqual(deadline, subscription.ackDeadline / 1000);
-
-        setImmediate(done);
-
-        return Promise.resolve();
-      };
-
-      subscription.renewLeases_();
-    });
-
-    it('should re-set the lease timeout', function(done) {
-      subscription.setLeaseTimeout_ = done;
-      subscription.renewLeases_();
     });
   });
 
@@ -1864,154 +709,6 @@ describe('Subscription', function() {
       };
 
       subscription.seek(FAKE_SNAPSHOT_NAME, gaxOpts, done);
-    });
-  });
-
-  describe('setFlushTimeout_', function() {
-    it('should set a flush timeout', function(done) {
-      var flushed = false;
-
-      subscription.flushQueues_ = function() {
-        flushed = true;
-      };
-
-      var delayPromise = delay(0);
-      var fakeBoundDelay = function() {};
-
-      delayPromise.cancel.bind = function(context) {
-        assert.strictEqual(context, delayPromise);
-        return fakeBoundDelay;
-      };
-
-      delayOverride = function(timeout) {
-        assert.strictEqual(timeout, 1000);
-        return delayPromise;
-      };
-
-      var promise = subscription.setFlushTimeout_();
-
-      promise.then(function() {
-        assert.strictEqual(subscription.flushTimeoutHandle_, promise);
-        assert.strictEqual(promise.cancel, fakeBoundDelay);
-        assert.strictEqual(flushed, true);
-        done();
-      });
-    });
-
-    it('should swallow cancel errors', function() {
-      var promise = subscription.setFlushTimeout_();
-      promise.cancel();
-      return promise;
-    });
-
-    it('should return the cached timeout', function() {
-      var fakeHandle = {};
-
-      subscription.flushTimeoutHandle_ = fakeHandle;
-
-      var promise = subscription.setFlushTimeout_();
-      assert.strictEqual(fakeHandle, promise);
-    });
-  });
-
-  describe('setLeaseTimeout_', function() {
-    var fakeTimeoutHandle = 1234;
-    var fakeRandom = 2;
-
-    var globalSetTimeout;
-    var globalMathRandom;
-
-    before(function() {
-      globalSetTimeout = global.setTimeout;
-      globalMathRandom = global.Math.random;
-    });
-
-    beforeEach(function() {
-      subscription.isOpen = true;
-      subscription.latency_ = {
-        percentile: function() {
-          return 0;
-        },
-      };
-    });
-
-    after(function() {
-      global.setTimeout = globalSetTimeout;
-      global.Math.random = globalMathRandom;
-    });
-
-    it('should set a timeout to call renewLeases_', function(done) {
-      var ackDeadline = (subscription.ackDeadline = 1000);
-
-      global.Math.random = function() {
-        return fakeRandom;
-      };
-
-      global.setTimeout = function(callback, duration) {
-        assert.strictEqual(duration, fakeRandom * ackDeadline * 0.9);
-        setImmediate(callback); // the done fn
-        return fakeTimeoutHandle;
-      };
-
-      subscription.renewLeases_ = done;
-      subscription.setLeaseTimeout_();
-      assert.strictEqual(subscription.leaseTimeoutHandle_, fakeTimeoutHandle);
-    });
-
-    it('should subtract the estimated latency', function(done) {
-      var latency = 1;
-
-      subscription.latency_.percentile = function(percentile) {
-        assert.strictEqual(percentile, 99);
-        return latency;
-      };
-
-      var ackDeadline = (subscription.ackDeadline = 1000);
-
-      global.Math.random = function() {
-        return fakeRandom;
-      };
-
-      global.setTimeout = function(callback, duration) {
-        assert.strictEqual(duration, fakeRandom * ackDeadline * 0.9 - latency);
-        done();
-      };
-
-      subscription.setLeaseTimeout_();
-    });
-
-    it('should not set a timeout if one already exists', function() {
-      subscription.renewLeases_ = function() {
-        throw new Error('Should not be called.');
-      };
-
-      global.Math.random = function() {
-        throw new Error('Should not be called.');
-      };
-
-      global.setTimeout = function() {
-        throw new Error('Should not be called.');
-      };
-
-      subscription.leaseTimeoutHandle_ = fakeTimeoutHandle;
-      subscription.setLeaseTimeout_();
-    });
-
-    it('should not set a timeout if the sub is closed', function() {
-      subscription.renewLeases_ = function() {
-        throw new Error('Should not be called.');
-      };
-
-      global.Math.random = function() {
-        throw new Error('Should not be called.');
-      };
-
-      global.setTimeout = function() {
-        throw new Error('Should not be called.');
-      };
-
-      subscription.isOpen = false;
-      subscription.setLeaseTimeout_();
     });
   });
 
@@ -2079,75 +776,6 @@ describe('Subscription', function() {
       };
 
       subscription.snapshot(SNAPSHOT_NAME);
-    });
-  });
-
-  describe('writeTo_', function() {
-    var CONNECTION_ID = 'abc';
-    var CONNECTION = {};
-
-    beforeEach(function() {
-      subscription.connectionPool = {
-        acquire: function(connId, cb) {
-          cb(null, CONNECTION);
-        },
-      };
-    });
-
-    it('should return a promise', function() {
-      subscription.connectionPool.acquire = function() {};
-
-      var returnValue = subscription.writeTo_();
-      assert(returnValue instanceof Promise);
-    });
-
-    it('should reject the promise if unable to acquire stream', function() {
-      var fakeError = new Error('err');
-
-      subscription.connectionPool.acquire = function(connId, cb) {
-        assert.strictEqual(connId, CONNECTION_ID);
-        cb(fakeError);
-      };
-
-      return subscription.writeTo_(CONNECTION_ID, {}).then(
-        function() {
-          throw new Error('Should not resolve.');
-        },
-        function(err) {
-          assert.strictEqual(err, fakeError);
-        }
-      );
-    });
-
-    it('should write to the stream', function(done) {
-      var fakeData = {a: 'b'};
-
-      CONNECTION.write = function(data) {
-        assert.strictEqual(data, fakeData);
-        done();
-      };
-
-      subscription.writeTo_(CONNECTION_ID, fakeData);
-    });
-
-    it('should capture the write latency when successful', function() {
-      var fakeLatency = 500;
-      var capturedLatency;
-
-      CONNECTION.write = function(data, cb) {
-        setTimeout(cb, fakeLatency, null);
-      };
-
-      subscription.latency_.add = function(value) {
-        capturedLatency = value;
-      };
-
-      return subscription.writeTo_(CONNECTION_ID, {}).then(function() {
-        var upper = fakeLatency + 50;
-        var lower = fakeLatency - 50;
-
-        assert(capturedLatency > lower && capturedLatency < upper);
-      });
     });
   });
 });
